fix: add Home page adn fix getSearch
