--- src/App.css
@@ -1,14 +1,0 @@
-#root {
-  max-width: 1380px;
-  margin: 0 auto;
-  padding: 0;
-}
-
-.main {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  gap: 20px;
-  padding: 40px;
-}
