+++ src/index.css
@@ -12,10 +12,16 @@
   --primary: #659c96;
   --secondary: #0f222b;
   --light: #a9bbba;
-  --background-color: #021b2794;
+  --background-color: #0f222bcc;
   --accent: #d5a900;
   --transition: all 0.3s ease-in-out;
   --shadow: 0px 2px 10px 0px #01141d;
+}
+
+#root {
+  max-width: 1380px;
+  margin: 0 auto;
+  padding: 0;
 }
 
 * {
