+++ src/main.tsx
@@ -2,14 +2,11 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App.tsx";
-import ErrorBoundary from "./components/ErrorBoundary/ErrorBoundary.tsx";
 
 const mountElement = document.getElementById("root") as HTMLElement;
 const root = createRoot(mountElement);
 root.render(
   <StrictMode>
-    <ErrorBoundary>
-      <App />
-    </ErrorBoundary>
+    <App />
   </StrictMode>,
 );
