+++ src/pages/HomePage/HomePage.module.css
@@ -1,0 +1,7 @@
+.main {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 40px;
+  padding: 40px;
+}
