+++ src/pages/HomePage/HomePage.props.ts
@@ -1,0 +1,9 @@
+import { PlanetParams } from "../../types/types";
+
+interface HomePageState {
+  items: PlanetParams[];
+  searchValue: string | null;
+  isLoading: boolean;
+}
+
+export default HomePageState;
